% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Count.clonal.mutations.R
\name{count.clonal.mutations}
\alias{count.clonal.mutations}
\title{Count clonal mutations on one or several chromosomal copies}
\usage{
count.clonal.mutations(
  aceseq.file.name,
  snv.file.name,
  purity.refit = F,
  chromosomes = c(1:22),
  max.CN = 4
)
}
\arguments{
\item{aceseq.file.name}{full path to the "comb_pro_extra" file as generated by ACEseq.}

\item{snv.file.name}{full path to the SNV file as generated by the DKFZ SNV calling pipeline.}

\item{purity.refit}{should the purity be re-estimated based on the VAF distribution? Defaults to FALSE}

\item{max.CN}{highest copy number to take into account. Must not exceed 6.}

\item{vector}{of chromosomes to be considered. Defaults to 1-22.}
}
\value{
a list containing with purity and the overall ploidy of the tumor, clonal.mutation.matrix: a matrix containing the number of clonal mutations per chromosome and state. Rows correspond to copy number/B-allele state, columns correspond to chromosome. Segment.length.matrix: length of the segments as in clonal.mutation.matrix. Copy.number.indicator: the copy number per state (corresponds to rows of clonal.mutation.matrix). B.allele.indicator: the B-allele count per state (corresponds to rows of clonal.mutation.matrix)
count.clonal.mutations()
}
\description{
This function counts the number of clonal mutations residing on a single or multiple copies per genomic fragment. Segments of equal copy number and B-allele count are merged per chromosome.
}
