% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Extract.copy.number.info.per.SSNV.R
\name{Extract.copy.number.info.per.SSNV}
\alias{Extract.copy.number.info.per.SSNV}
\title{Extract the copy number information per SSNV}
\usage{
Extract.copy.number.info.per.SSNV(vcf, cnv)
}
\arguments{
\item{vcf}{Mutation file in VCF representation as a list (as returned by read.vcf from package bedR).}

\item{cnv}{ACEseq file as data frame}
}
\value{
a list containing the coverage ratio, B-allele frequency, genotype, total copy number and A allele count per mutation
Extract.copy.number.info.per.SSNV()
}
\description{
This function takes as input a vcf file and assigns each mutation a copy number state based on the output by ACEseq.
}
