% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Extract.read.counts.from.vcf.R
\name{Extract.info.from.vcf}
\alias{Extract.info.from.vcf}
\title{Extracts information from a vcf file}
\usage{
Extract.info.from.vcf(
  vcf,
  info = "readcounts",
  type = "snvs",
  mutationcaller = "DKFZ"
)
}
\arguments{
\item{vcf}{Mutation file in VCF representation as a list (as returned by read.vcf from package bedR).}

\item{info}{Which information is to be retrieved? Possible values are "readcounts", "depth", "VAF", "AA_change". Defaults to "readcounts".}

\item{type}{Is the VCF file reporting snvs or indels? Possible values are "snvs" and "indel". Defaults to "snvs".}

\item{mutationcaller}{which mutation caller was used to generate the vcf-files? Defaults to "DKFZ".}
}
\value{
purity and ploidy.
Extract.info.from.vcf()
}
\description{
Extracts information from a vcf file
}
