% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Extract.purity.ploidy.from.ACEseq.R
\name{Extract.purity.ploidy.from.ACEseq}
\alias{Extract.purity.ploidy.from.ACEseq}
\title{Extracts purity and ploidy estimates from ACEseq file name}
\usage{
Extract.purity.ploidy.from.ACEseq(aceseq.filename)
}
\arguments{
\item{aceseq.filename}{The name of the ACEseq file.}
}
\value{
purity and ploidy.
Extract.purity.ploidy.from.ACEseq()
}
\description{
Extracts purity and ploidy estimates from ACEseq file name
}
