% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRCA_ECA_quantification.R
\name{MRCA.ECA.quantification}
\alias{MRCA.ECA.quantification}
\title{Estimate mutation counts at MRCA and ECA.}
\usage{
MRCA.ECA.quantification(
  clonal.mutation.matrix.per.tumor,
  segment.length.matrix.per.tumor,
  chromosomes = c(1:22),
  haploid.genome.length = 3.3 * 10^9,
  minimal.segment.size = 10^7,
  max.CN = 4
)
}
\arguments{
\item{clonal.mutation.matrix.per.tumor}{Matrix with mutations sorted by copy number and allele as returned by Count.clonal.mutations; rows: CN/B state, columns: chromosomes}

\item{segment.length.matrix.per.tumor}{Matrix with segment lengths sorted by copy number and allele as returned by Count.clonal.mutations; rows: CN/B state, columns: chromosomes}

\item{chromosomes}{chromosomes of interest. Defaults to 1-22.}

\item{haploid.genome.length}{size of the genome under consideration. Defaults to 3.3*10^9 (human genome)}

\item{minimal.segment.size}{What is the minimal segment size to take into consideration? Defaults to 10^7bp.}

\item{max.CN}{maximal copy number to be considered. Defaults to 4.  Which information is to be retrieved? Possible values are "readcounts", "depth", "VAF", "AA_change". Defaults to "readcounts".}

\item{mutationcaller}{which mutation caller was used to generate the vcf-files? Defaults to "DKFZ".}
}
\value{
A list containing the estimates for the mutation time at MRCA and ECA along with lower and upper boundaries according to 95\% confidence intervals estimated by bootstrapping the segments 1000 times. Earliest.mutation.times returns the density of events earlier than a common ECA. Other list elements report the segments assigned to ECA and MRCA: gains.at.earliest.time - chromosomal gains assigned to the earliest datable time point (NA if all early gains map to a common ECA), gains.not.mapping.to.earliest.time - gains that do not map to MRCA, ECA or the earliest time point, other.gains.mapping.to.earliest.time - gains that can be mapped to the earliest time point or to ECA or to MRCA, gains.at.mrca - gains mapping to the MRCA (may also include gains that can also be mapped to ECA, i.e., not uniquely dateable gains. These pop up in gains.at.mrca.conforming.eca), gains.at.mrca.conforming.eca - gains mapping to the ECA or the MRCA, gains.uniquely.mapped.to.eca - gains mapping to ECA only, gains.not.maping.to.eca.or.mrca - gains that cannot be mapped to the ECA and cannot be mapped to the MRCA, monosomic.states.not.matching.mrca - monosomic states (i.e., non-amplified clonal segments) that do not conform to the MRCA,
MRCA.ECA.quantification()
}
\description{
Estimate mutation counts at MRCA and ECA.
}
