% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mutation.time.converter.R
\name{Mutation.time.converter}
\alias{Mutation.time.converter}
\title{Convert mutation counts into mutation times}
\usage{
Mutation.time.converter(clonal.mutations, max.CN = 4)
}
\arguments{
\item{clonal.mutations}{A vector with clonal mutation counts for different copy number and allele states. Corresponds to a column of the clonal.mutation.matrix as returned by Count.clonal.mutations().}

\item{max.CN}{maximal copy number to be considered. Defaults to 4.  Which information is to be retrieved? Possible values are "readcounts", "depth", "VAF", "AA_change". Defaults to "readcounts".
Mutation.time.converter()}
}
\description{
The number of clonal mutations found in a clonal peak does not necessarily correspond to "mutation time" as mutations are lost from the non-amplified clonal peak by a chromosomal gain. This function corrects for such losses, rendering mutation times comparable.
}
