% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model_NB_initiation.R
\name{P.1st.driver}
\alias{P.1st.driver}
\title{Probability to acquire the first hit at time t. Only written for the contraction case}
\usage{
P.1st.driver(parms, t)
}
\arguments{
\item{parms}{a list containing delta1 (loss rate during expansion), delta2 (loss rate during contraction, set to arbitrary value if modeling contraction), N (log10 of maximal precursor count), mu1 (rate of acquiring the first oncogenic event, log10), mu (neutral mutation rate), r (selective advantage associated with first oncogenic event)}

\item{t}{time
P.1st.driver()}
}
\description{
Probability to acquire the first hit at time t. Only written for the contraction case
}
