% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model_NB_initiation.R
\name{P.2nd.driver.1st.hit.during.expansion}
\alias{P.2nd.driver.1st.hit.during.expansion}
\title{Probability to acquire the first hit during exponential expansion and the second hit during contraction}
\usage{
P.2nd.driver.1st.hit.during.expansion(
  lambda1,
  delta1,
  N,
  t,
  lambda2,
  delta2 = NA,
  p.surv,
  mu1,
  mu2,
  r,
  model = "contraction"
)
}
\arguments{
\item{lambda1}{proliferation rate during expansion; set to 1}

\item{delta1}{loss rate during expansion}

\item{N}{maximal precursor count}

\item{t}{time}

\item{lambda2}{proliferation rate during contraction; set to 1}

\item{delta2}{loss rate during contraction, set to arbitrary value if modeling homeostasis rather than contraction}

\item{p.surv}{survival probability of 2nd mutant}

\item{mu1}{rate of acquiring 1st oncogenic event}

\item{mu2}{rate of acquiring 2nd oncogenic event}

\item{r}{selective advantage associated with 1st oncogenic event}

\item{model}{inidicates whether expansion and homeostasis or expansion and contraction is modeled
P.2nd.driver.1st.hit.during.expansion()}
}
\description{
Probability to acquire the first hit during exponential expansion and the second hit during contraction
}
