% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model_NB_initiation.R
\name{P.2nd.driver}
\alias{P.2nd.driver}
\title{Probability to acquire two hits at time t}
\usage{
P.2nd.driver(parms, t, model = "contraction")
}
\arguments{
\item{parms}{a list containing delta1 (loss rate during expansion), delta2 (loss rate during contraction, set to arbitrary value if modeling homeostasis), N (log10 of maximal precursor count), p.surv (survival probability associated with 2nd event), mu1 (rate of acquiring the first oncogenic event, log10), mu2 (rate of acquiring the second oncogenic event, log10), mu (neutral mutation rate), r (selective advantage associated with first oncogenic event)}

\item{t}{time}

\item{model}{inidicates whether expansion and homeostasis or expansion and contraction is modeled
P.2nd.driver()}
}
\description{
Probability to acquire two hits at time t
}
