% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model_NB_initiation.R
\name{P.both.hits.during.expansion}
\alias{P.both.hits.during.expansion}
\title{Probability to acquire 2 mutations during exponential expansion (compare Haeno et al., 2007)}
\usage{
P.both.hits.during.expansion(lambda1, delta1, N, t, lambda2, mu1, mu2, p.surv)
}
\arguments{
\item{lambda1}{proliferation rate during expansion; set to 1}

\item{delta1}{loss rate during expansion}

\item{N}{maximal precursor count}

\item{t}{time}

\item{lambda2}{proliferation rate during contraction; set to 1}

\item{mu1}{rate of acquiring 1st oncogenic event}

\item{mu2}{rate of acquiring 2nd oncogenic event}

\item{p.surv}{survival probability of 2nd mutant
P.both.hits.during.expansion()}
}
\description{
Probability to acquire 2 mutations during exponential expansion (compare Haeno et al., 2007)
}
