% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model_NB_initiation.R
\name{P.eca.at.t}
\alias{P.eca.at.t}
\title{Probability of the first event happening at time t1 given the second event happened at t2}
\usage{
P.eca.at.t(parms, t1, t2, model = "contraction")
}
\arguments{
\item{parms}{a list containing delta1 (loss rate during expansion), delta2 (loss rate during contraction), N (log10 of maximal precursor count), p.surv (survival probability associated with 2nd event), mu1 (rate of acquiring the first oncogenic event, log10), mu2 (rate of acquiring the second oncogenic event, log10), mu (neutral mutation rate), r (selective advantage associated with first oncogenic event)}

\item{t1}{time point of first mutation}

\item{t2}{time point of second mutation}

\item{model}{inidicates whether expansion and homeostasis or expansion and contraction is modeled
P.eca.at.t()}
}
\description{
Probability of the first event happening at time t1 given the second event happened at t2
}
