% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model_NB_initiation.R
\name{Sample.timepoint}
\alias{Sample.timepoint}
\title{Sample the time point of tumor initiation}
\usage{
Sample.timepoint(parms, model = "contraction")
}
\arguments{
\item{parms}{a list containing delta1 (loss rate during expansion), delta2 (loss rate during contraction), N (log10 of maximal precursor count), p.surv (survival probability associated with 2nd event), mu1 (rate of acquiring the first oncogenic event, log10), mu2 (rate of acquiring the second oncogenic event, log10), mu (neutral mutation rate), r (selective advantage associated with first oncogenic event)}

\item{model}{inidicates whether expansion and homeostasis or expansion and contraction is modeled
Sample.timepoint()}
}
\description{
Sample the time point of tumor initiation
}
